+++ WebApp/src/index.ts
@@ -20,7 +20,7 @@ export class RenderStreaming {
           .option('-s, --secure', 'Enable HTTPS (you need server.key and server.cert)', process.env.SECURE || false)
           .option('-k, --keyfile <path>', 'https key file (default server.key)', process.env.KEYFILE || 'server.key')
           .option('-c, --certfile <path>', 'https cert file (default server.cert)', process.env.CERTFILE || 'server.cert')
-          .option('-w, --websocket', 'Enable Websocket Signaling', process.env.WEBSOCKET || false)
+          .option('-t, --type <type>', 'Type of signaling protocol, Choose websocket or http (default websocket)', process.env.TYPE || 'websocket')
           .option('-m, --mode <type>', 'Choose Communication mode public or private (default public)', process.env.MODE || 'public')
           .option('-l, --logging <type>', 'Choose http logging type combined, dev, short, tiny or none.(default dev)', process.env.LOGGING || 'dev')
           .parse(argv);
@@ -30,7 +30,7 @@ export class RenderStreaming {
           secure: option.secure == undefined ? false : option.secure,
           keyfile: option.keyfile,
           certfile: option.certfile,
-          websocket: option.websocket == undefined ? false : option.websocket,
+          type: option.type == undefined ? 'websocket' : option.type,
           mode: option.mode,
           logging: option.logging,
         };
@@ -69,9 +69,17 @@ export class RenderStreaming {
         }
       });
     }
+    if (this.options.type == 'http') {
+      console.log(`Use http polling for signaling server.`);
+    }
+    else if(this.options.type != 'websocket') {
+      console.log(`signaling type should be set "websocket" or "http". ${this.options.type} is not supported.`);
+      console.log(`Changing signaling type to websocket.`);
+      this.options.type = 'websocket';
+    }
+    if (this.options.type == 'websocket') {
+      console.log(`Use websocket for signaling server ws://${this.getIPAddress()[0]}`);
 
-    if (this.options.websocket) {
-      console.log(`start websocket signaling server ws://${this.getIPAddress()[0]}`);
       //Start Websocket Signaling server
       new WSSignaling(this.server, this.options.mode);
     }